feat: ComputeEngineCredential supports specifying scopes
The existing behavior of ComputeEngineCredentials is to not provide the
`?scopes` parameter in its access token request to the metadata
server.  This is workable on GCE, which has the concept of a scope
lock / default scopes applied to the VM.  However, neither new
AppEngine nor GKE Workload Identity have this concept; they just
return a hardcoded set of default scopes if no specific scopes are
requested.  This makes usage of some Google APIs impossible using the
Java client libraries.

This change adds scope support to ComputeEngineCredentials.  It's
optional, so the current behavior will stay the default.  However, if
the user specifies scopes on the builder or using `createScoped`, then
those scopes will be passed to the metadata server.