Override AwsSecurityCredentials
See https://issuetracker.google.com/issues/238911014

When using a Web Identity token file (`AWS_WEB_IDENTITY_TOKEN_FILE`) to authenticate with AWS,
environment variables `AWS_ACCESS_KEY_ID`, `AWS_SECRET_ACCESS_KEY`, and `AWS_SESSION_TOKEN` are not set.

While the existing implementation then makes a request to Amazon's EC2
metadata service endpoint; when the Java instance is running in a pod in
EKS, the metadata endpoint does not return metadata for the pod, but for
the underlying EKS node.

To work around this, we can optionally set these variables in the
credential builder instead of looking up the env variables, or polling
the metadata endpoint.